Inline the core libs into the central module and de-duplicate
Move the chord identification (chordlib), speech rendering (speech) and
chord reporting (chordreport) code out of their own files into the central
timtam_AccessiChords.lua module, so all core code lives in one place. The
three files are removed and the move-to-chord actions now go through
AccessiChords.reportChord() / moveToChord(). README updated accordingly.

Also collapse duplication that came in alongside the new modules:
- merge the duplicate OSARA-output helpers say()/speak() into one guarded speak()
- drop the dead chordlib.intervals / chordlib.tensionDegree tables (no readers;
  speech.interval / speech.tension are the live equivalents)
- extract chordModeRange(), shared by playNotesByChordMode and
  insertMidiNotesByChordMode
- extract AccessiChords.moveToChord() so the move-to-next/previous-chord action
  scripts (which must stay as separate REAPER actions) are thin callers of one
  shared helper

Behaviour verified identical to the previous three-file version via an offline
Lua harness: chord reporting, chord insert/preview order, and the move-chord
actions all produce identical output.
